+++ languagetool-core/src/main/java/org/languagetool/rules/Category.java
@@ -18,11 +18,11 @@
  */
 package org.languagetool.rules;
 
+import org.jetbrains.annotations.Nullable;
+
 /**
  * A rule's category. Categories are used to group rules for
  * a better overview.
- * 
- * @author Daniel Naber
  */
 public final class Category {
 
@@ -35,19 +35,25 @@ public final class Category {
   }
 
   private final String name;
+  private final CategoryId id;
   private final Location location;
   private final boolean defaultOff;
 
   /**
-   * Create a new category.
+   * @since 3.3
+   */
+  public Category(CategoryId id, String name) {
+    this(id, name, Location.INTERNAL, true);
+  }
+
+  /**
    * @param name name of the category
    */
-  public Category(final String name) {
+  public Category(String name) {
     this(name, Location.INTERNAL);
   }
 
   /**
-   * Create a new category.
    * @param name name of the category
    * @since 2.8
    */
@@ -61,9 +67,26 @@ public final class Category {
    * @since 2.9
    */
   public Category(String name, Location location, boolean onByDefault) {
+    this.id = null;
     this.name = name;
     this.location = location;
     this.defaultOff = !onByDefault;
+  }
+
+  /**
+   * @since 3.3
+   */
+  public Category(CategoryId id, String name, Location location, boolean onByDefault) {
+    this.id = id;
+    this.name = name;
+    this.location = location;
+    this.defaultOff = !onByDefault;
+  }
+
+  /** @since 3.3 */
+  @Nullable
+  public CategoryId getId() {
+    return id;
   }
 
   public String getName() {