+++ languagetool-core/src/main/java/org/languagetool/rules/Rule.java
@@ -216,6 +216,7 @@ public abstract class Rule {
   }
 
   public final void setCategory(final Category category) {
+    Objects.requireNonNull(category, "category cannot be null");
     this.category = category;
   }