+++ languagetool-language-modules/de/src/main/java/org/languagetool/rules/de/MissingVerbRule.java
@@ -58,7 +58,7 @@ public class MissingVerbRule extends GermanRule {
     rule2 = new PatternRule("internal", language, Arrays.asList(
             new PatternToken("Herzlichen", true, false, false),
             new PatternToken("Glückwunsch", true, false, false)), "", "", "");
-    super.setCategory(new Category(messages.getString("category_grammar")));
+    super.setCategory(new Category(CategoryIds.GRAMMAR, messages.getString("category_grammar")));
     setDefaultOff();
     addExamplePair(Example.wrong("<marker>In diesem Satz kein Wort.</marker>"),
                    Example.fixed("In diesem Satz <marker>fehlt</marker> kein Wort."));